+++ core/proguard-rules.pro
@@ -29,9 +29,9 @@
     @retrofit2.http.* <methods>;
 }
 
--keep public class core.mvp.activities.AbstractActivity {*;}
--keep public class core.mvp.presenters.AbstractPresenter {*;}
--keep public class core.mvp.views.AbstractView {*;}
+-keep public class com.weekendwars.core.mvp.activities.AbstractActivity {*;}
+-keep public class com.weekendwars.core.mvp.presenters.AbstractPresenter {*;}
+-keep public class com.weekendwars.core.mvp.views.AbstractView {*;}
 
 # Ignore annotation used for build tooling.
 -dontwarn org.codehaus.mojo.animal_sniffer.IgnoreJRERequirement
