+++ core/src/main/java/com/weekendwars/core/mvp/views/AbstractView.java
@@ -1,0 +1,4 @@
+package com.weekendwars.core.mvp.views;
+
+public interface AbstractView {
+}
