--- core/src/main/java/core/mvp/views/AbstractView.java
@@ -1,4 +1,0 @@
-package core.mvp.views;
-
-public interface AbstractView {
-}
