+++ sampleapp/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="core.mvp.sampleapp">
+    package="com.weekendwars.core.mvp.sampleapp">
 
     <application
         android:allowBackup="true"
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="core.mvp.sampleapp.activities.TestActivity">
+        <activity android:name="com.weekendwars.core.mvp.sampleapp.activities.TestActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
