+++ sampleapp/src/main/java/com/weekendwars/core/mvp/sampleapp/views/TestView.java
@@ -1,0 +1,7 @@
+package com.weekendwars.core.mvp.sampleapp.views;
+
+import com.weekendwars.core.mvp.views.AbstractView;
+
+public interface TestView extends AbstractView {
+    void displayRestoredState(long savedState);
+}
