--- sampleapp/src/main/java/core/mvp/sampleapp/views/TestView.java
@@ -1,7 +1,0 @@
-package core.mvp.sampleapp.views;
-
-import core.mvp.views.AbstractView;
-
-public interface TestView extends AbstractView {
-    void displayRestoredState(long savedState);
-}
